Task0

+++ Tyuiu.KomarovaMV.Sprint5.Task0.V5.Lib/DataService.cs
@@ -7,7 +7,7 @@ namespace Tyuiu.KomarovaMV.Sprint5.Task0.V5.Lib
         public string SaveToFileTextData(int x)
         {
             string path = Path.GetTempFileName();
-            double z = Math.Round(Math.Log((x+1)/(x+2)),3);
+            double z = Math.Round(Math.Log(((double)x + 1) / ((double)x + 2)), 3);
             File.WriteAllText(path,Convert.ToString(z));
             return path;
         }
